Implement same logic for `startUpload`

+++ services/alp-dataflow/src/prefect/prefect.service.ts
@@ -235,6 +235,10 @@ export class PrefectService {
       }
 
       await this.prefectExecutionClient.executePipInstall(userId, modifiedFileStem)
+      if (existingFlowMetadata) {
+        await this.prefectFlowService.deleteFlowMetadata(existingFlowMetadata.flowId)
+      }
+      await this.prefectFlowService.createFlowMetadata(flowMetadataInput)
     } catch (err) {
       if (!existingFlowMetadata) {
         await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
@@ -247,11 +251,6 @@ export class PrefectService {
         await this.prefectFlowService.updateDefaultPluginStatus(defaultPluginId, PluginUploadStatus.FAILED)
       }
       throw new InternalServerErrorException(errorMessage)
-    } finally {
-      if (existingFlowMetadata) {
-        await this.prefectFlowService.deleteFlowMetadata(existingFlowMetadata.flowId)
-      }
-      await this.prefectFlowService.createFlowMetadata(flowMetadataInput)
     }
 
     try {
@@ -339,8 +338,18 @@ export class PrefectService {
     this.logger.info(`Deployment Folder: ${deploymentFolderPath}`)
 
     let flowMetadataInput
-    let newFlowMetadata
+    let existingFlowMetadata
     try {
+      // prepare metadata input
+      if (defaultDeploymentInfo && defaultDeploymentInfo.url) {
+        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, defaultDeploymentInfo.url)
+      } else {
+        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, url)
+      }
+
+      // if metadata with same flowId exists
+      existingFlowMetadata = await this.prefectFlowService.getFlowMetadataById(flowMetadataInput.flowId)
+
       defaultDeploymentInfo.defaultPluginId
       if (defaultPluginId) {
         await this.prefectFlowService.updateDefaultPluginStatus(defaultPluginId, PluginUploadStatus.INSTALLING)
@@ -364,25 +373,14 @@ export class PrefectService {
       }
 
       await this.prefectExecutionClient.executePipInstall(userId, modifiedFileStem)
-
-      // prepare metadata input
-      if (defaultDeploymentInfo && defaultDeploymentInfo.url) {
-        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, defaultDeploymentInfo.url)
-      } else {
-        flowMetadataInput = await this.prepareFlowMetadata(deploymentFolderPath, url)
+      if (existingFlowMetadata) {
+        await this.prefectFlowService.deleteFlowMetadata(existingFlowMetadata.flowId)
       }
-
-      // if metadata with same flowId exists
-      const existFlowMetadata = await this.prefectFlowService.getFlowMetadataById(flowMetadataInput.flowId)
-      if (existFlowMetadata) {
-        await this.prefectFlowService.deleteFlowMetadata(existFlowMetadata.flowId)
-      }
-      newFlowMetadata = await this.prefectFlowService.createFlowMetadata(flowMetadataInput)
     } catch (err) {
-      if (newFlowMetadata) {
+      if (!existingFlowMetadata) {
         await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
+        await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
       }
-      await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
       this.deleteDeploymentFolder(deploymentFolderPath)
       const errorMessage = `Error installing pip package, check if package is valid`
       this.logger.error(`${errorMessage}: ${err}`)
@@ -407,8 +405,10 @@ export class PrefectService {
         await this.prefectFlowService.updateDefaultPluginStatus(defaultPluginId, PluginUploadStatus.COMPLETE)
       }
     } catch (err) {
-      await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
-      await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
+      if (!existingFlowMetadata) {
+        await this.prefectFlowService.deleteFlowMetadata(flowMetadataInput.flowId)
+        await this.prefectApi.deleteFlow(flowMetadataInput.flowId)
+      }
       const errorMessage = `Error creating flow with file ${fileStem}${fileType}`
       this.logger.error(`${errorMessage}: ${err}`)
       if (defaultPluginId) {
